120. Triangle
